requirements.txt
docker_py==1.10.6
PyYAML==6.0
requests==2.27.1
